refactor: 재사용을 위해 아이템 컴포넌트 분리
