+++ src/pages/cart/components/CartList.tsx
@@ -1,7 +1,10 @@
 import CartItem from './CartItem';
 import AllSelectCheckBox from './AllSelectCheckBox';
-import { allCartItemStates } from '@/store/atoms';
-import { useRecoilValue } from 'recoil';
+import { useCartContext } from '../context/CartContext';
+import CheckAndDelete from './CheckAndDelete';
+import CartItemImageAndInfo from './CartItemImageAndInfo';
+import CountButton from './CountButton';
+import styles from './cartItemImageAndInfo.module.css';
 
 export default function CartList() {
   const cartItems = useRecoilValue(allCartItemStates);
@@ -12,15 +15,17 @@ export default function CartList() {
       <ul>
         {cartItems.map((cartItem) => {
           return (
-            <CartItem
-              key={cartItem.id}
-              id={cartItem.id}
-              name={cartItem.product.name}
-              price={cartItem.product.price}
-              category={cartItem.product.category}
-              imageUrl={cartItem.product.imageUrl}
-              isChecked={cartItem.product.isChecked}
-            />
+            <CartItem key={cartItem.id}>
+              <CheckAndDelete id={cartItem.id} />
+              <CartItemImageAndInfo
+                className={styles.itemImageAndInfoContainer}
+                imageUrl={cartItem.product.imageUrl}
+                name={cartItem.product.name}
+                price={cartItem.product.price}
+              >
+                <CountButton id={cartItem.id} />
+              </CartItemImageAndInfo>
+            </CartItem>
           );
         })}
       </ul>
